+++ CPAC/utils/monitoring/__init__.py
@@ -2,9 +2,11 @@
 
 See https://fcp-indi.github.com/docs/developer/nodes for C-PAC-specific documentation.
 See https://nipype.readthedocs.io/en/latest/api/generated/nipype.utils.profiler.html for Nipype's documentation.'''  # noqa E501
+from .custom_logging import set_up_logger
 from .monitoring import LoggingHTTPServer, LoggingRequestHandler, \
                         log_nodes_cb, log_nodes_initial, monitor_server, \
                         recurse_nodes
 
 __all__ = ['LoggingHTTPServer', 'LoggingRequestHandler', 'log_nodes_cb',
-           'log_nodes_initial', 'monitor_server', 'recurse_nodes']
+           'log_nodes_initial', 'monitor_server', 'recurse_nodes',
+           'set_up_logger']